pyproject.toml
[tool.poetry]
name = "vg-music-scraper"
version = "0.1.0"
description = "artifact to download videogames ost from khinsider page"
authors = ["chelosky <marcelolazo19972013@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
urllib3 = "^2.1.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
tqdm = "^4.66.1"
pathvalidate = "^3.2.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"

[tool.poetry.scripts]
start = "vg_music_scraper:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
